Move the FeatureForm Command implementation to shared library and update definition to match Maui implementation
