+++ src/Toolkit/Toolkit/UI/Controls/FeatureForm/FeatureFormView.cs
@@ -70,25 +70,6 @@ namespace Esri.ArcGISRuntime.Toolkit.UI.Controls
             Unloaded += (s, e) => ClearUtilityAssociationCandidateSelection();
         }
 
-        private class Command : System.Windows.Input.ICommand
-        {
-            private Action _execute;
-            private Func<bool> _canExecute;
-
-            public Command(Action execute, Func<bool> canExecute)
-            {
-                _execute = execute;
-                _canExecute = canExecute;
-            }
-            internal void RaiseCanExecuteChanged() => CanExecuteChanged?.Invoke(this, EventArgs.Empty);
-
-            public event EventHandler? CanExecuteChanged;
-
-            public bool CanExecute(object? parameter) => _canExecute();
-
-            public void Execute(object? parameter) => _execute();
-        }
-
         /// <summary>
         /// Command for calling <see cref="FinishEditingAsync(bool)"/> and applying edits to the currently active Feature Form if all fields are valid.
         /// </summary>
@@ -176,7 +157,7 @@ namespace Esri.ArcGISRuntime.Toolkit.UI.Controls
             foreach (var item in GetDescendentsOfType<FieldFormElementView>(this))
             {
                 item.ResetValidationState();
-                ((Command)FinishEditingCommand).RaiseCanExecuteChanged();
+                ((Command)FinishEditingCommand).ChangeCanExecute();
             }
         }
 
@@ -252,8 +233,8 @@ namespace Esri.ArcGISRuntime.Toolkit.UI.Controls
             {
                 this.Dispatch(() =>
                 {
-                    ((Command)FinishEditingCommand).RaiseCanExecuteChanged();
-                    ((Command)DiscardEditsCommand).RaiseCanExecuteChanged();
+                    ((Command)FinishEditingCommand).ChangeCanExecute();
+                    ((Command)DiscardEditsCommand).ChangeCanExecute();
                 });
             }
         }
@@ -276,7 +257,7 @@ namespace Esri.ArcGISRuntime.Toolkit.UI.Controls
 #if MAUI
                     base.OnPropertyChanged(nameof(IsValid));
 #endif
-                    ((Command)FinishEditingCommand).RaiseCanExecuteChanged();
+                    ((Command)FinishEditingCommand).ChangeCanExecute();
                 }
             }
         }
@@ -495,8 +476,7 @@ namespace Esri.ArcGISRuntime.Toolkit.UI.Controls
                 {
                     item.ResetValidationState();
                 }
-                ((Command)FinishEditingCommand).RaiseCanExecuteChanged();
-
+                ((Command)FinishEditingCommand).ChangeCanExecute();
                 if (requireAllErrorsResolved && ScrollToFirstError()) return false;
                 await FinishEditingAsync();
                 return true;
@@ -1047,8 +1027,8 @@ namespace Esri.ArcGISRuntime.Toolkit.UI.Controls
                 inpcNew.PropertyChanged += _elementPropertyChangedListener.OnEvent;
             }
             UpdateIsValidProperty();
-            ((Command)FinishEditingCommand).RaiseCanExecuteChanged();
-            ((Command)DiscardEditsCommand).RaiseCanExecuteChanged();
+            ((Command)FinishEditingCommand).ChangeCanExecute();
+            ((Command)DiscardEditsCommand).ChangeCanExecute();
         }
 
         /// <summary>
